pyproject.toml
[project]
name = "limpopo-mega-similarity"
version = "0.1.0"
description = """     
    1) Communicating with the Business Agent to check for any new content.
    2) Vectorizing the new content, transforming it into numerical vectors.
    3) Performing a similarity search using the vectorized content, matching it against user profiles or other content for recommendations or categorization.
"""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "fastapi==0.115.6",
    "filelock==3.16.1",
    "fsspec==2024.10.0",
    "grpcio-tools==1.68.1",
    "grpcio==1.68.1",
    "mysql-connector-python==9.1.0",
    "python-dotenv==1.0.1",
    "qdrant-client==1.12.1",
    "requests==2.32.3",
    "sentence-transformers==3.3.1",
    "uvicorn==0.33.0",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.26.5",
    "hyperframe==6.0.1",
    "idna==3.10",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.0",
    "packaging==24.2",
    "pillow==11.0.0",
    "portalocker==2.10.1",
    "protobuf==5.29.1",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "safetensors==0.4.5",
    "scikit-learn==1.6.0",
    "scipy==1.14.1",
    "setuptools==75.6.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "sympy==1.13.1",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "torch==2.5.1",
    "tqdm==4.67.1",
    "transformers==4.47.0",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "pytest>=8.3.4",
]
